# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STWV Multi-Company Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /compliance:
    get:
      operationId: listCompliance
      tags: [compliance]
      summary: List compliance checklist items
      security: [{ bearerAuth: [] }]
      parameters:
        - name: companyId
          in: query
          required: false
          schema:
            type: integer
        - name: regime
          in: query
          required: false
          schema:
            type: string
            description: "germany or india"
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of compliance items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceItem"
    post:
      operationId: createComplianceItem
      tags: [compliance]
      summary: Create a compliance checklist item
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplianceItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
  /compliance/{id}:
    patch:
      operationId: updateComplianceItem
      tags: [compliance]
      summary: Update a compliance item status/notes
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplianceItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceItem"
    delete:
      operationId: deleteComplianceItem
      tags: [compliance]
      summary: Delete a compliance item
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /compliance/seed:
    post:
      operationId: seedCompliance
      tags: [compliance]
      summary: Seed compliance items for a company/regime/year
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedComplianceBody"
      responses:
        "201":
          description: Items seeded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplianceItem"

  # ───── AUDIT LOGS ─────
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit_logs]
      summary: List audit log entries (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: entity_type
          schema: { type: string }
        - in: query
          name: entity_id
          schema: { type: integer }
        - in: query
          name: actor_id
          schema: { type: integer }
        - in: query
          name: action
          schema: { type: string }
        - in: query
          name: date_from
          schema: { type: string }
        - in: query
          name: date_to
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        "200":
          description: List of audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
  /projects/{id}/activity:
    get:
      operationId: getProjectActivity
      tags: [audit_logs]
      summary: Get project-scoped activity feed (last 50 events)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Project activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # ───── ADMIN DASHBOARD ─────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserBasic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - email
        - role
...
    ComplianceItem:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        regime:
          type: string
          description: "One of: germany, india"
        year:
          type: integer
        quarter:
          type: ["integer", "null"]
        month:
          type: ["integer", "null"]
        itemKey:
          type: string
        itemLabel:
          type: string
        deadline:
          type: string
        status:
          type: string
          description: "One of: pending, filed, overdue"
        responsibleUserId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        filedAt:
          type: ["string", "null"]
          format: date-time
        responsibleUser:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/UserBasic"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - regime
        - year
        - itemKey
        - itemLabel
        - deadline
        - status
        - createdAt
        - updatedAt
    CreateComplianceItemBody:
      type: object
      properties:
        companyId:
          type: integer
        regime:
          type: string
        year:
          type: integer
        quarter:
          type: ["integer", "null"]
        month:
          type: ["integer", "null"]
        itemKey:
          type: string
        itemLabel:
          type: string
        deadline:
          type: string
        status:
          type: string
        responsibleUserId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - companyId
        - regime
        - year
        - itemKey
        - itemLabel
        - deadline
    UpdateComplianceItemBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]
        responsibleUserId:
          type: ["integer", "null"]
        deadline:
          type: string
        filedAt:
          type: ["string", "null"]
    SeedComplianceBody:
      type: object
      properties:
        companyId:
          type: integer
        regime:
          type: string
          description: "One of: germany, india"
        year:
          type: integer
      required:
        - companyId
        - regime
        - year

    # ─── Admin Dashboard ───
...
    AuditLogActor:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
    AuditLog:
      type: object
      required:
        - id
        - createdAt
        - actorRole
        - action
        - entityType
        - entityId
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        actorId:
          type: ["integer", "null"]
        actorRole:
          type: string
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        entityLabel:
          type: ["string", "null"]
        oldValue:
          type: ["object", "null"]
          additionalProperties: true
        newValue:
          type: ["object", "null"]
          additionalProperties: true
        projectId:
          type: ["integer", "null"]
        actor:
          oneOf:
            - $ref: "#/components/schemas/AuditLogActor"
            - type: "null"